requirements.txt
# Core
torch>=2.1.0
torch-geometric>=2.4.0
torch-scatter
torch-sparse
torch-cluster

# Medical imaging / mesh processing
nibabel>=5.0.0
scikit-image>=0.21.0
trimesh>=4.0.0
scipy>=1.11.0

# ML utilities
numpy>=1.24.0
scikit-learn>=1.3.0
pyyaml>=6.0

# Visualisation
matplotlib>=3.7.0
pandas>=2.0.0

# Optional (latent space visualisation)
umap-learn>=0.5.0

# Development
jupyter
ipykernel
